+++ NEWS.md
@@ -1,3 +1,4 @@
 # lorax (development version)
 
+* `extract_rules()` gains an `lgb.Booster` method for extracting rules from LightGBM models.
 * `extract_rules()` gains an `xgb.Booster` method for extracting rules from XGBoost models.